+++ include/hal_io.h
@@ -5,50 +5,21 @@
  *
  * ## Capabilities
  *
- * - Pin direction can be set to input or output
- *   - Pull-up or pull-down selection when pin is configured as input
- * - High or low signal output when pin is configured as output
+ * - Pin direction can be set to input or output via hal_io_configure()
+ *   - Pull-up or pull-down selection when pin is configured as input via \ref
+ *     hal_io_pin_configuration's `is_pull_up` member
+ * - High or low signal output when pin is configured as output via
+ *   hal_io_write()
  *   - Toggle pin output signal when pin is configured as output
- * - Pin level reading when pin is configured as input
+ * - Pin state reading when pin is configured as input via hal_io_read()
  *
  * ## Function Return Type
  *
- * Every I/O function will return \ref io_result. This value can be checked if
- * operation was successful or wasn't.
- *
- * ## Configuring Pins
- *
- * Configuring a I/O pin is done with io_configure() function. This function
- * will accept any kind of information about a pin.
- *
- * ### Switching Between Input and Output
- *
- * \ref io_pin_configuration's `direction` member can be used to specify
- * direction of an I/O pin.
- *
- * ### Setting Pin As Pull-Up/Pull-Down In Input Mode
- *
- * I/O pins with input direction, can be configured with pull-up enabled or
- * disabled. To do this, \ref io_pin_configuration's `is_pull_up` member can be
- * used.
- *
- * ## Read/Write Operations On Pins
- *
- * ### Change Pin Output State
- *
- * I/O pins can be driven high or low with io_write() function.
- *
- * ### Toggling Pin Output State
- *
- * Pin state can be toggled with io_toggle() function.
- *
- * ### Reading Pin Input State
- *
- * Pin state can be read with io_read() function. Read value should be stored
- * in given \ref io_pin_state pointer.
+ * Every I/O function will return \ref hal_result_io. This value can be checked
+ * if operation were successful or weren't.
  * */
 
-// SPDX-FileCopyrightText: 2023 Ceyhun Şen <ceyhuusen@gmail.com>
+// SPDX-FileCopyrightText: 2025 Ceyhun Şen <ceyhuusen@gmail.com>
 // SPDX-License-Identifier: MIT
 
 #ifndef __HAL_IO_H
@@ -57,19 +28,22 @@
 #include <stdint.h>
 
 /**
- * @enum io_result
+ * @enum hal_result_io
  * @brief Module specific return results.
  */
-enum io_result {
-    io_success = 0,
-    io_error,
+enum hal_result_io {
+    hal_result_io_ok = 0,
+    hal_result_io_error_invalid_pin,
+    hal_result_io_error_invalid_port,
+    hal_result_io_error_invalid_direction,
+    hal_result_io_error_invalid_state
 };
 
 /**
- * @enum io_port
+ * @enum hal_io_port
  * @brief Port specifier of a I/O pin.
  */
-enum io_port { io_port_b = 0, io_port_c, io_port_d };
+enum hal_io_port { hal_io_port_b = 0, hal_io_port_c, hal_io_port_d };
 
 /**
  * I/O pin specifier.
@@ -77,22 +51,22 @@ enum io_port { io_port_b = 0, io_port_c, io_port_d };
  * @param port Target port.
  * @param pin Target pin of the port
  */
-struct io_pin {
-    enum io_port port;
+struct hal_io_pin {
+    enum hal_io_port port;
     uint8_t pin;
 };
 
 /**
- * @enum io_pin_state
+ * @enum hal_io_pin_state
  * @brief State of the current I/O pin.
  */
-enum io_pin_state { io_state_low = 0, io_state_high = 1 };
+enum hal_io_pin_state { hal_io_state_low = 0, hal_io_state_high = 1 };
 
 /**
- * @enum io_pin_direction
+ * @enum hal_io_pin_direction
  * @brief Direction settings for I/O pins.
  */
-enum io_pin_direction { io_direction_output, io_direction_input };
+enum hal_io_pin_direction { hal_io_direction_output, hal_io_direction_input };
 
 /**
  * Configuration options for an I/O pin.
@@ -101,15 +75,18 @@ enum io_pin_direction { io_direction_output, io_direction_input };
  * @param pull_up Is pull-up enabled? 0 on disabled, 1 (or other values) on
  * enabled.
  */
-struct io_pin_configuration {
-    enum io_pin_direction direction;
+struct hal_io_pin_configuration {
+    enum hal_io_pin_direction direction;
     uint8_t is_pull_up;
 };
 
-enum io_result io_configure(struct io_pin io,
-                            struct io_pin_configuration configuration);
-enum io_result io_write(struct io_pin io, enum io_pin_state state);
-enum io_result io_toggle(struct io_pin io);
-enum io_result io_read(struct io_pin io, enum io_pin_state *state);
+enum hal_result_io
+hal_io_configure(struct hal_io_pin io,
+                 struct hal_io_pin_configuration configuration);
+enum hal_result_io hal_io_write(struct hal_io_pin io,
+                                enum hal_io_pin_state state);
+enum hal_result_io hal_io_toggle(struct hal_io_pin io);
+enum hal_result_io hal_io_read(struct hal_io_pin io,
+                               enum hal_io_pin_state *state);
 
 #endif // __HAL_IO_H