+++ CHANGELOG.md
@@ -6,6 +6,16 @@ also be mirrored to the Github releases page.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- I/O module re-enabled.
+
+### Changed
+
+- I/O module naming conventions.
+
 ## [0.4.1]
 
 ### Added