+++ CHANGELOG.md
@@ -6,11 +6,13 @@ also be mirrored to the Github releases page.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.0] - TODO
 
 ### Added
 
 - I/O module re-enabled.
+- Timer0
+  - Set output compare mode
 
 ### Changed
 
@@ -104,7 +106,8 @@ branch.
   - TWI
   - ADC
 
-[unreleased]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.4.1...HEAD
+[unreleased]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.6.0...HEAD
+[0.5.0]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.3...v0.4.0
 [0.3]: https://github.com/ceyhunsen/ATmega328P-HAL-Driver/compare/v0.2...v0.3